[FIX] User: Time Slots before 10:00 AM
A bug in displaying the timeslots on Haircut-Book page was fixed.

The bug was due to the unchecked time format of the javascipt time function at 08:00 AM. It should be in the format of '08:00:00', but the actual format of the time was '8:00:00'. So, just fixed that with appending a '0' in front of the hour if the hour is less than 10.